Add linting (flake8, pylint) to tox, CI processes
This adds flake8 and pylint steps to the CI pipeline. For pylint,
because the output messages are so comprehensive, I've split this into
two separate tox commands: `pylint_critical` limits output to errors and
warnings only, while `pylint_all` outputs everything. The former just
checks for critical problems and is used in CI to prevent builds from
happening if it doesn't pass. The latter can be used on an as-needed
basis to check for possible problems and "code smells" -- since it isn't
built into the CI process, we can choose which issues to address and
which to ignore.

Next commit(s) will be resolving the critical issues that flake8/pylint
identify.