Clean up tests based on pylint suggestions
This adjusts some of the things that pylint calls out when running with
all notifications turned on and makes a handful of other adjustments.

The last test in `test_data.py` had an incorrect assertion that was
passing when it should sometimes have been raising an error --
basically, `assert [...]` instead of `assert all([...])`. Fixing this
now causes this test to fail intermittently.