Fix precise_distribution_counter method
The previous commit uncovered an issue with the test that tested this
method, where one aspect of the method under test wasn't being tested
correctly (and was always passing) due to a syntax error. Fixing the
syntax error caused the test to (sometimes) fail, and I found that one
of the assumptions about how the method under test worked was incorrect.

I ended up massively refactoring that method, but it now works more
closely to what I originally intended. I modified tests related to that
method so they are now more accurate.