Move code from old data module; remove
This moves the code that was in solrfixtures/data.py into separate files
in the new data directory and hooks things up via data/__init__.py so
that everything still works and tests still pass even though we've
removed data.py. In future commits we'll be refactoring some of this
further.