Eliminate RNG use for options => a static value
For the IntEmitter, if the supplied min/max values are equal, then
there's no need to make a random selection -- just return that static
value, whatever it is. For the StrEmitter, if the alphabet is one
character or element, there's no need to make a random choice -- just
use that one value.