Add a Chance (e.g. percent chance) emitter
This is a more concrete implementation of the True/False Choice emitter
needed for profile.Fields. (Used to determine whether or not to output a
value. So -- in a schema, a given field might only be populated 30% of
the time. In that case your field could use a Chance gate emitter set at
30%.