Generalize EmitterGroup (ObjectGroup)
This changes EmitterGroup to ObjectGroup, as the EmitterGroup
implementation already worked for any kind of object. Now that I'm
implementing a new Schema class, I'd like to use this for groups of
fields, and it's clear I needed to generalize EmitterGroup this way.
Additionally, ObjectGroup is now a subclass of UserList and behaves like
a list.