Add resetting the unique shuffle to Choice.seed
I think, calling `seed` on any of the RandomEmitters implies that the
new seed is applied immediately and you can begin emitting values using
the new seed without taking further steps. For Choice emitters where
`unique` is True, global uniqueness is implemented using a static
shuffle -- items are shuffled once and then returned in shuffle order.
In that case, in order to apply the seed, the shuffle has to be reset,
which loses track of what values were already emitted. I *think* that is
okay -- it seems incredibly unlikely that someone would need to reseed
after using the emitter and NOT expect it to reshuffle. This keeps the
`seed` behavior consistent, which seems better than forcing people to
know they need to `reset` after reseeding only in certain cases.