Make small change to improve Choice emitter
Incidentally, I realized I could make a tiny improvement to the Choice
emitter's `emit` algorithm: when 'each_unique' is True, meaning each
call to `emit` must produce a unique list of values, but values can be
reused each time -- if only one value is requested at a time, then it's
functionally no different than an emitter without the 'each_unique'
constraint. Without the 'each_unique' constraint, the selection
algorithm is *much* faster, especially if you also have weights. So I
added an additional condition so that it only uses the slower algorithm
when 'each_unique' is True if the number of values requested is >1.

This only affects Choice emitters where 'each_unique' is True and one
value is requested at a time, but the difference is substantial.

This change did require a change to the test, even though we're using
seed values. The change in the selection algorithm ends up picking
different values, although it still makes sense given the provided
weights.