Add emitters.wrappers and Wrap emitter
This adds an emitter to let us wrap other emitters to modify their
output without having to proliferate Emitter classes. This will be most
useful for doing data conversions. (Still need to add tests.)

Not sure if it would be useful to provide options for adding arguments
to pass to the wrapper.