Create 'fixed' emitter category
Just some moving around and renaming:
- Created emitters.fixed module.
- Moved emitter.StaticEmitter to emitters.fixed.Static.
- Moved emitters.counters.Sequential to emitters.fixed.Iterative.

The idea being that `fixed` emitters output relatively fixed data,
rather than random data. (I know, the emitter categories so far are not
great.)

Also, StaticEmitter is an implementation (not a base class) so it never
really made sense to put it in the 'emitter' module along with Emitter
and RandomEmitter.