Improve uniqueness properties on text emitters
One of the things that I kind of want to be able to do but cannot find a
good implementation for that works well in all situations is emitting
unique randomized words with the text.Word emitter. This doesn't add
that capability, but it improves the ability to calculate the number of
possible unique values that a given text.Word or text.Text emitter can
emit. I also updated some of the doc strings here that were either wrong
or missing or just out of date.