Create ItemsMixin, for emitter `items` interface
Some emitters have a finite set of values that they can emit. Previously
I haphazardly implemented a similar interface for many (but not all) of
these, using an `items` attribute. I've found that using a mixin for
this is helpful, to provide consistent behavior.
- `items` is immutable, implemented via a @property with no setter.
- The `num_unique_values` property by default returns a number
  reflecting the set of unique values in self.items -- e.g.,
  len(set(self.items)).

This commit adds the mixin and implements it in all the emitter classes
that can possibly use it.