Fix num_unique_values being None for Text emitter
... when the `sep_emitter` attribute is None. Previously, without
bothering to check, I forgot that this can actually be None -- when it
is, a ' ' is used as the separator character when values are generated.
When I added the `_update_num_unique_vals` method, I was adding a check
for presence of this emitter and skipping the whole calculation when it
was not there. On top of that, I neglected to test for that eventuality
so missed it in tests. Long story short, this adds a couple of tests for
it and fixes the bug. If the `sep_emitter` is None, then the calculation
assumes that emitter contributes exactly 1 unique value.