Add ability to 'hide' fields
This is the first step in adding more features for basing field output
off other fields' output.

Sometimes you may need a field that generates some sort of intermediary
value but isn't directly included in the schema, where you will have
other fields that use the output from this field. To make that work,
I've added a 'hide' attribute to the Field class. When True, the field
output is generated and cached like always but the Schema removes it
from the final output dictionary.