Add emitters.fromfields.BasedOnFields
This adds a `BasedOnFields` emitter type, a companion to `CopyFields`
that lets you emit some value or values that are *based on* the values
output by the source field or fields. The user supplies the list of
source fields and an 'action' function that receives the source output
and does something with it to produce the final output.

Note that the same results were already possible using a wrapper from
`emitters.wrappers` to wrap a CopyFields emitter. However, the
BasedOnFields emitter is faster, particularly when you need to source
more than one field. Plus it has built-in handling for RNG that's
consistent with other RNG handling, which the wrappers currently lack.