Add ChildrenMixin and RandomWithChildrenMixin
ChildrenMixin is a mixin for creating compound emitters, or any object
that needs to manage one or more attached child emitters. On
initialization, you simply pass in a mapping (dict) of labels to
child emitters (as the 'children' kwarg), and then you gain access to an
'emitters' ObjectMap property where you can access child emitters by
label. Also, calling the `reset` method resets all children.

RandomWithChildrenMixin combines RandomMixin and ChildrenMixin, for any
object that needs BOTH children emitters and RNG (or where any children
need RNG). In addition to giving you both RandomMixin and ChildrenMixin
features, it also ensures that children are seeded when you call `seed`
or `reset` on the parent. (I.e., this combined behavior would not be
present if you just used RandomMixin and ChildrenMixin, which is why the
combined mixin is required.)