Fix the Iterative and Sequential emitters
The previous implementation of the Iterative emitter was flawed; it was
possible for a repeatedly reset iterator to lose state somehow and begin
returning None. In revisiting the code, I realized that the previously
complex implementation could be replaced by a simple infinite iterator
that just continuously yields items from `self.iterator_factory`, using
a simple `while: True` loop. The kicker is that this doesn't work if you
supply an iterator_factory that returns an empty iterator, such as
`lambda: iter([])`. In this case, the infinite iterator that gets
generated never yields anything and never calls StopIteration, which
causes an infinite loop. To address this, I added a check when assigning
the iterator_factory to ensure it isn't empty (it raises a ValueError if
it is).

Because the Sequential emitter is a simple subclass of Iterative, this
fixes the same issue with that.