Rename everything to fauxdoc; add useful README
Next steps in getting this ready for the first public release. I'm
renaming the entire project from `solrfixtures` to `fauxdoc`. It started
out as a way to generate fixtures for Solr but then became a
general-purpose package for generating fake document/record data in
general.

This also adds a more helpful README. (Still need some work on this, but
it's a start.)