Remove need for Poetry in development
Now that setuptools allows for editable installs without a `setup.py`
file, for this project there's not really an advantage to using Poetry
in development. Removing Poetry removes a potential barrier for
contributing and simplifies things a bit. It also lets us use a
`pyproject.toml` that conforms to PEP 621.