Update publish.yml: run manually or on tags
Still trying to work out a `publish` workflow that works well. With this:
- Pushing to a branch no longer triggers publication.
- Pushing a tag should always trigger publication.
- Upload to Test PyPI always runs when triggered. (I.e., all tagged
  versions go to Test PyPI.)
- Upload to live PyPI only happens on a stable public release: version
  string should not contain `dev`, `a` (alpha), `b` (beta), or `rc`
  (release candidate).
- OR: the workflow can be triggered manually, from any branch or tag.
  (The live PyPI filter still applies.)

After tagging and publishing, we can then manually create releases on
GitHub and reference the appropriate tag.

In the future, we could add a step that would automatically create a
GitHub release for us after publishing to live PyPI. OR, if we want to
predicate publishing to live PyPI on creating a release, we could set
`on.release` and modify `on.workflow_run.tags` so the auto-publication
only triggers for non-stable releases.