Finish adding the `publish to live PyPI` step
Looks like everything is working, so I think I'm ready to try a live
upload to PyPI. Previously I had it simply echoing a message if it
reached that point so it wouldn't accidentally upload something to live
PyPI before I was ready. This adds the action that will actually publish
it there.