Fix wrapper emitters
This contains a major update to emitters.wrappers, resulting from
ironing out all of the type-checking issues. Turns out implementing
type-checking for flexible, somewhat dynamic callbacks (like wrapped
functions) is kind of a nightmare. I realized that my previous
implementation of a generic `Wrap` base class, with subclasses `WrapOne`
and `WrapMany` that each inherit from `Wrap` but allow different
signatures for the wrapped function, probably violates the Liskov
substitution principle. I reimplemented this by converting `Wrap` into a
`Wrapper` class, which is no longer a base class but is instead used for
the wrapper objects (instantiated with the user-provided wrapper
function) that implement the needed behavior and are members of each of
`WrapOne` and `WrapMany`. I think this solves the Liskov violation and
makes the code for `WrapOne` and `WrapMany` a lot cleaner, too.

`Wrap` is no longer needed, but, since it's part of the public API
release with v1.0.0, I've left it in. It will be deprecated in a future
version.

The other big change here is that `Wrapper` allows for trying a mock
call to the underlying wrapper function using inspect.signature.bind. So
now, instantiating one of the Wrap* emitters uses this to test whether
the user-supplied wrapper function has the expected signature, instead
of waiting until an actual call is attempted. (I wanted to use type
checking for this, but I need to be able to use ParamSpec, which mypy
doesn't yet fully implement -- specifically, allowing partial signatures
and using Concatenate.)

The public API should be the same, except for the addition of the
Wrapper class. I did have to update a handful of tests, but mainly just
the ones related to checking for errors when calling the wrapper and how
those errors are reported.

Type hints have also been updated so that mypy tests for this module
pass.