Enforce correct overloading of Emitter __call__
I finally figured out how to use the typing.overload decorator correctly
to indicate that calling an emitter like `emitter()` should return one
value and `emitter(1)` should output a list of values. This solves
several mypy errors where the emitter call was correct in a given
context, but mypy couldn't tell which of the two was actually expected
from context alone.

(The trick was that I needed to add decorated hints both to the
EmitterLike protocol AND to base Emitter class. Previously I had only
added it to the EmitterLike protocol. And, in that case, if I tried to
use any kind of Emitter as an EmitterLike, it would complain that the
call signatures didn't match.)