Fix type hints for `fromfields` emitters
Like with wrappers, addressing the problems with type hints for
`fromfields` emitters revealed some issues with the code architecture.
I tried to address these without changing the API.

The biggest issue was that BasedOnFields subclassed CopyFields when it
shouldn't have -- it just needed to borrow some of the implementation
details around source fields, but it wasn't a proper subtype or
subclass. To fix this, I abstracted out the source field functionality
and created a separate object type (SourceFieldGroup). CopyFields and
BasedOnFields are now sibling classes, not parent/child.

I also simplified BasedOnFields by implementing the `action` callable
using the `wrappers.BoundWrapper` type. It had pretty much the same
behavior already, but this just formalizes that.

Like with wrappers, I had to update the fromfields tests because of the
change to error checking the `action` call signature -- now that gets
checked when the emitter is initialized instead of later when it's
called.