Fix `profile` type hints
This was pretty straightforward, except I did slightly change how the
Field child emitters are handled: they are stored directly as instance
attributes, and then an `_emitters` ObjectMap attribute is created that
maps the attribute objects. (Instead of creating the ObjectMap first and
then using properties to map those values to attributes.) This is much
simpler and works better for typing but still allows the ObjectMap to
behave as intended.