Fix type hints for `text` emitters
Mostly straightforward changes here. I made similar changes to the
emitters here re: children that I made to `profile.Field` children --
instead of creating a "children" ObjectMap and then creating properties
that that map each child emitter to an attribute, now we store the
applicable emitters directly as instance attributes and create the
ObjectMap to reference the attribute objects. This is functionally
equivalent but a little simpler and works better for typing.