Ensure properties w/setters expect the right type
While working on the last commit (updates to wrappers), I ran across a
type-hinting/mypy issue, discussed at length here:
https://github.com/python/mypy/issues/3004.

The gist is, when you have a property that uses a setter, and when you
set that property (i.e. with `foo.bar = 'value'`) -- mypy expects the
value you're setting to be whatever type the getter returns, regardless
of what type the setter takes. If the setter takes e.g. a broader type
and then normalizes it so that the getter returns a more specific type,
then mypy will report a type error when you attempt to set the attribute
using the broader type.

Since I use this in a few places, my solution is to set the type hint
for the property getter so that it reflects the broader type -- whatever
type the setter takes. That way it is still technically correct and
avoids the mypy error, although it is not as helpful as it could be if
it reflected the narrow type that the property actually stores/returns.

This commit updates that wherever it occurs.