Add deprecation warnings for deprecated components
This adds standardized deprecation warnings for things that have already
been deprecated and slated for removal in v2. Where necessary I've added
a __getattr__ method at the module level so that warnings are issued
when a deprecated module attribute is imported or accessed, whether
it's a variable, class, etc.

These specific things have been deprecated and now issue a warning when
accessed or used:
- typing.Number: Use `float`.
- typing.EmitterLikeCallable: Use `EmitterLike`.
- typing.StrEmitterLike: Use `EmitterLike[str]`.
- typing.IntEmitterLike: Use `EmitterLike[int]`.
- typing.BoolEmitterLike: Use `EmitterLike[bool]`.
- emitters.wrappers.Wrap: Use `emitters.wrappers.WrapOne` or `.WrapMany`.
- emitters.Wrap: Same as previous.
- emitters.fixed.Sequential.iterator_factory: Being able to set this
  attribute is deprecated; in the future it will be read-only. If you
  need a different sequence, create a new Sequential instance.

I've also added tests for the deprecation warnings, which bumps the
minimum required pytest version to 3.8.0 for Python 3.7, 3.8, and 3.9 --
pytest did not handle these correctly in earlier versions.