Fix setting `source` and `action`, and types
For the `source` attribute on CopyFields/BasedOnFields and the `action`
attribute on BasedOnFields: when these attributes are set by a user,
either on or after initialization, the most reasonable type of value for
the user to provide is not the type of value that gets stored. For
`source`, the user provides a Field instance or a list of Field
instances, and these then get wrapped in a `SourceFieldGroup` instance.
For `action`, the user provides a function or other callable and this
gets wrapped in a `fauxdoc.emitters.wrappers.BoundWrapper` instance.

Instead of using a shortcut where a property setter does this conversion
on assignment, I've added new `set` methods that do this:
`CopyFields.set_source_fields`, `BasedOnFields.set_source_fields`, and
`BasedOnFields.set_action_function`. This behavior is (arguably) clearer
and more expected -- assigning a value to an attribute directly and then
accessing that attribute will return the same value. The new methods
make it clearer that you're setting from a different type. This also
works better for type checking.

Also in this commit, I've changed `BasedOnFields.action` to a property,
which now does the appropriate validation checks when the attribute is
set.