Fix setting `wrapper`, and types
For the `wrapper` attribute on WrapOne/WrapMany: when a user sets this,
either on or after initialization, the most reasonable type of value
they would provide is a function. But the value that actually gets
stored is a `BoundWrapper` instance that wraps their function.

To make this seamless, I had been using a shortcut where the property
setter does the conversion on assignment -- however, the resulting
behavior is (perhaps) a little counterintuitive. It's clearer, and it
works better for type checking, if the attribute stores the value as
assigned. If we want to do a type conversion, we should instead use a
separate method to set the attribute from a different user-provided
type. So, I've added `set_wrapper_function` for this: the user provides
a function or other callable; it converts this to a BoundWrapper
instance and assigns it to the `wrapper` attribute.

This is consistent with the other similar changes I've just made to
`fauxdoc.emitters.fromfields`.