Fix setting `Schema.fields` and types
This is similar to the last two commits. For the `fields` attribute on
Schema: when a user sets this, whether on or after initialization, the
most reasonable type of value they would provide is a Field instance or
a list of Field instances. (In fact, to make it even easier, the
existing `__init__` and `add_fields` method accept a star argument:
e.g., schema.add_fields(field1, field2, field3).) But -- the value that
actually gets stored is an `ObjectMap` instance wrapping the provided
fields.

To make this seamless, I had been using a shortcut where the property
setter does the conversion on assignment -- however, the resulting
behavior is (perhaps) a little counterintuitive, since the attribute
stores a different value than what was assigned. It also somewhat
confounds type checkers. If we want to do a type conversion, using a
separate method is probably the clearest way. In this case, we already
have the `add_fields` method; I've just added a `reset` kwarg, which, if
True, sets the attribute from scratch using the provided fields instead
of adding them to the existing fields. The default is False so that
existing `add_fields` calls still work as before.