Allow users to supply Choice.cum_weights
I've gone back and forth on whether Choice.cum_weights should be
settable or read-only; I've decided to go all the way to the "fully
settable" side, mainly because `random.choices` allows you to supply
either weights or cum_weights. The fauxdoc v1.0.0 API accidentally
allowed cum_weights to be settable; originally I planned to force it to
be read-only in v1.1.0, but now I think this will work better.

I added all the needed functionality for this:
- `cum_weights` is now an optional kwarg for __init__.
- If both `weights` and `cum_weights` are supplied to __init__, it
  raises a TypeError.
- When one gets set, the other gets updated appropriately.
- Checks/validation that were applied when setting `weights` are applied
  when setting `cum_weights`.
- Tests have been updated to thoroughly test all of the above.