Change `add_fields` w/reset=True to `set_fields`
On further reflection, changing the behavior of the existing
`fauxdox.profile.Field.add_fields` method so that it can also serve as a
setter is probably a bad idea. I've restored the original `add_fields`
behavior and instead implemented a new `set_fields` method. (This is so
that the `fields` attribute, which wants an ObjectMap, doesn't have to
be set directly using an ObjectMap, which is inconvenient.)